% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dBPGC.R
\name{dBPGC}
\alias{dBPGC}
\title{Calculate the Probability Density Function for the Bivariate Poisson-Gamma Distribution}
\usage{
dBPGC(x, y, params)
}
\arguments{
\item{x}{integer values (discrete variable)}

\item{y}{Numeric value (continuous variable)}

\item{params}{A numeric vector of parameters: \code{c(m10, m01, m11, m02, m12)}.}
}
\value{
A numeric value or vector representing the PDF evaluated at the given \code{x} and \code{y}.
}
\description{
Computes the probability density function (PDF) for the bivariate Poisson-Gamma distribution given parameters and values of X and Y.
}
\details{
This function calculates the PDF for the bivariate Poisson-Gamma distribution based on the provided parameters.
The PDF is given by:
\deqn{f(x, y; \mathbf{m}) = \frac{\exp(c + m10 \cdot x - m01 \cdot y - m11 \cdot x \cdot y + m02 \cdot \log(y) + m12 \cdot x \cdot \log(y))}{x! \cdot y}}
where \code{c} is a normalization constant calculated using \code{calC}.
}
\examples{
# Example usage:
params <- c(0.1, 0.2, 0.3, 0.4, 0.5)
x <- 5
y <- 2
dBPGC(x, y, params)

}
