% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ePLOT.R
\name{ePLOT}
\alias{ePLOT}
\title{Empirical Plot with Bivariate Poisson-Gamma Fit}
\usage{
ePLOT(X, Y)
}
\arguments{
\item{X}{Numeric vector representing the Poisson-distributed variable.}

\item{Y}{Numeric vector representing the Gamma-distributed variable.}
}
\description{
Generates a 3D empirical plot of the data along with the estimated bivariate Poisson-Gamma distribution.
}
\details{
This function creates a 3D empirical plot using the given data points for the Poisson and Gamma-distributed variables.
It also fits a bivariate Poisson-Gamma distribution to the data and superimposes the estimated density on the empirical plot.
}
\examples{
# Example usage:
params <- c(1,5,0,5,0)
sim_data <- rBPGC(params, points = 1e6, seed = 42)
XX <- sim_data$x
YY <- sim_data$y
X <- rpois(100,exp(1))
Y <- rgamma(100, 5, 5)
ePLOT(X, Y)

}
