% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FFtest.R
\name{FFtest}
\alias{FFtest}
\title{Fasano-Franceschini Test for goodness-of-fit of BPGC distribution}
\usage{
FFtest(X, Y)
}
\arguments{
\item{x}{integer values (discrete variable)}

\item{y}{Numeric value (continuous variable)}
}
\value{
The result of the Fasano-Franceschini test, which includes a test statistic and a p-value indicating the likelihood of the null hypothesis.
}
\description{
This function performs the Fasano-Franceschini test in order to test goodness-of-fit of BPGC distrubution.
}
\examples{
\dontrun{
# Example usage
params <- c(0.1, 0.2, 0.3, 0.4, 0.5)
sim_data <- rBPGC(params, points = 100, seed = 42)
X <- sim_data$x
Y <- sim_data$y
FFtest(X, Y)
}

}
