% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hospitalData.R
\name{hospitalData}
\alias{hospitalData}
\title{Hospital Dataset}
\format{
A data frame with 500 rows and 3 variables:
\describe{
  \item{day}{An integer vector indicating the day in order (1, 2, ...).}
  \item{number_of_admissions}{An integer vector representing the number of admissions per day.}
  \item{total_cost_of_treatment}{A numeric vector representing the total cost of treatment.}
}
}
\source{
Simulated data
}
\description{
A toy example of a hospital dataset.
}
\examples{
data("hospitalData")
head(hospitalData)

}
