% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pBPGC.R
\name{pBPGC}
\alias{pBPGC}
\title{Calculate the Cumulative Distribution Function for the Bivariate Poisson-Gamma Distribution}
\usage{
pBPGC(X, Y, params)
}
\arguments{
\item{params}{A numeric vector of parameters: \code{c(m10, m01, m11, m02, m12)}.}

\item{x}{integer values (discrete variable)}

\item{y}{Numeric value (continuous variable)}
}
\value{
A numeric value representing the CDF evaluated at the given \code{x} and \code{y}.
}
\description{
Computes the cumulative distribution function (CDF) for the bivariate Poisson-Gamma distribution given parameters and values of X and Y.
}
\details{
This function calculates the CDF for the bivariate Poisson-Gamma distribution based on the provided parameters by integrating the PDF.
}
\examples{
# Example usage:
params <- c(1,5,1,5,1)
x <- 2
y <- 0.3
pBPGC(x, y, params)

}
