% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOT.R
\name{PLOT}
\alias{PLOT}
\title{Plot the Bivariate Poisson-Gamma Distribution and Marginal Densities}
\usage{
PLOT(X, Y, params)
}
\arguments{
\item{X}{Numeric vector representing the range of values for the Poisson-distributed variable.}

\item{Y}{Numeric vector representing the range of values for the Gamma-distributed variable.}

\item{params}{A numeric vector of parameters: \code{c(m10, m01, m11, m02, m12)}.}
}
\description{
Generates a 3D plot of the bivariate Poisson-Gamma distribution along with its marginal densities.
}
\details{
This function creates a 3D surface plot of the bivariate Poisson-Gamma distribution using the provided parameters.
It also includes the marginal densities for X and Y.
}
\examples{
# Example usage:
X <- seq(0, 5, by = 1)
Y <- seq(0.1, 3, length.out = 50)
params <- c(1,5,1,5,1)
PLOT(X, Y, params)

}
