% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.mlEstBPGC.R
\name{print.mlEstBPGC}
\alias{print.mlEstBPGC}
\title{Print method for BPGC maximum likelihood estimates}
\usage{
\method{print}{mlEstBPGC}(x, ...)
}
\arguments{
\item{x}{An object of class \code{mlEstBPGC}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print method for BPGC maximum likelihood estimates
}
