% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rBPGC.R
\name{rBPGC}
\alias{rBPGC}
\title{Simulate Data Using Gibbs Sampling}
\usage{
rBPGC(params, points = 500, seed = NULL)
}
\arguments{
\item{params}{A numeric vector of parameters: \code{c(m10, m01, m11, m02, m12)}.}

\item{points}{The number of data points to generate. Default is 500.}

\item{seed}{An optional integer seed for reproducibility. Default is \code{NULL}.}
}
\value{
A data frame with simulated values of X and Y.
}
\description{
Generates a dataset by simulating values of X and Y using Gibbs sampling.
}
\details{
This function generates a dataset using Gibbs sampling based on the provided parameters.
It starts with initial values and iteratively samples values of X and Y from their respective
conditional distributions until the specified number of points is generated.
}
\examples{
# Example usage:
params <- c(0.1, 0.2, 0.3, 0.4, 0.5)
sim_data <- rBPGC(params, points = 100, seed = 42)
head(sim_data)

}
