% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlEst.R
\name{mlEst}
\alias{mlEst}
\title{Maximum Likelihood Estimation (MLE) Function}
\usage{
mlEst(X, Y, params_init = rep(0.5, 5), max_iterations = 1000, rel_tol = 1e-08)
}
\arguments{
\item{X}{A numeric vector of observations for the predictor variable.}

\item{Y}{A numeric vector of observations for the response variable.}

\item{params_init}{Initial values for the parameters.}

\item{max_iterations}{Maximum number of iterations for optimization.}

\item{rel_tol}{Relative tolerance for convergence.}
}
\value{
A list with the following components:
\describe{
  \item{params}{Optimized parameters.}
  \item{negative_log_likelihood}{Value of the negative log-likelihood at the optimum.}
  \item{convergence}{Convergence code: 0 for successful convergence, non-zero otherwise.}
  \item{message}{Description of the optimization result.}
}
}
\description{
Estimates the parameters that maximize the likelihood of the data.
}
\details{
This function performs Maximum Likelihood Estimation (MLE) using constrained optimization
to find the parameters that maximize the likelihood of the data given the model. It uses
the \code{constrOptim} function for optimization, which requires setting constraints on
the parameters.

The negative log-likelihood function used for optimization is internally defined within
this function. It calculates the negative log-likelihood based on given parameters and
observed data.
}
\examples{
# Example usage:
params <- c( 1, 1, 0.1, 1, 0.1)
sim_data <- rBPGC(params, points = 1000000, seed = 42)
X <- sim_data$x
Y <- sim_data$y
mleEst(X, Y)

}
